Move avatar helpers out of application helper
Put them in an avatar helper specific class for clarity.
